streamelements integration, simulate events
