+++ SubathonManager.Core/Events/StreamElementsEvents.cs
@@ -1,0 +1,11 @@
+﻿namespace SubathonManager.Core.Events;
+
+public class StreamElementsEvents
+{
+    public static event Action<bool>? StreamElementsConnectionChanged;
+
+    public static void RaiseStreamElementsConnectionChanged(bool isConnected)
+    {
+        StreamElementsConnectionChanged?.Invoke(isConnected);
+    }
+}
