+++ SubathonManager.Integration/SubathonManager.Integration.csproj
@@ -7,6 +7,7 @@
     </PropertyGroup>
 
     <ItemGroup>
+      <PackageReference Include="StreamElementsNET" Version="1.0.0" />
       <PackageReference Include="TwitchLib.Api" Version="3.9.0" />
       <PackageReference Include="TwitchLib.Client" Version="3.4.0" />
       <PackageReference Include="TwitchLib.Communication" Version="1.0.6" />
